refactor(go-template): extract ctor lowering into memo/ctor-lowering.ts
Move the constructor-context expression lowering out of the adapter
class into pure free functions (Roadmap B): lowerCtorExpr, lowerCtorCond,
lowerCtorStringArray. They lower the narrow expression surface a
derived-state memo needs (#1897) into Go constructor code, reading only
state.localConstants / state.propsObjectName / state.needsStringsImport
and parseLiteralExpression.

No new GoEmitContext member is needed; the two external call sites call
lowerCtorExpr(this.emitCtx, …).

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
